+++ apps/client/src/app/services/data.service.ts
@@ -150,6 +150,10 @@ export class DataService {
     return this.http.delete<any>(`/api/v1/order/${aId}`);
   }
 
+  public deleteAllOrders() {
+    return this.http.delete<any>(`/api/v1/order/`);
+  }
+
   public deleteUser(aId: string) {
     return this.http.delete<any>(`/api/v1/user/${aId}`);
   }