+++ libs/ui/src/lib/activities-table/activities-table.component.ts
@@ -52,6 +52,7 @@ export class ActivitiesTableComponent implements OnChanges, OnDestroy, OnInit {
   @Output() export = new EventEmitter<string[]>();
   @Output() exportDrafts = new EventEmitter<string[]>();
   @Output() import = new EventEmitter<void>();
+  @Output() deleteAllActivities = new EventEmitter<void>();
   @Output() importDividends = new EventEmitter<UniqueAsset>();
   @Output() selectedActivities = new EventEmitter<Activity[]>();
 
@@ -229,6 +230,10 @@ export class ActivitiesTableComponent implements OnChanges, OnDestroy, OnInit {
           return activity.id;
         })
     );
+  }
+
+  public onDeleteAllActivities() {
+    this.deleteAllActivities.emit();
   }
 
   public onImport() {